rewrote first test to push data to a sqlite database dataframes are now a fixture

+++ tests/integration_test/test_integration_test.py
@@ -1,21 +1,16 @@
 from logging import RootLogger
 import pytest
 
+# TODO: we will add a different test for data_collection_to_df
 from gpu4bdgamers.data_coll_script import (
     data_collection_to_df, data_collection_to_db
 )
 from gpu4bdgamers.database import push_to_db, TodayDataAlreadyExistsError
 
-import os
-import dotenv
-
 import pandas as pd
 import datetime
 import sqlalchemy
 
-
-dotenv.load_dotenv()
-test_db_url = os.getenv("test_db_url")
 
 def sql_query_format(table_name:str):
     """formats the sql query so we can just plug in table_name 
@@ -30,18 +25,14 @@ def delete_db_today_rows(
 ) -> None:
     """deletes all rows with today's date in a database table"""
     today = datetime.datetime.today().strftime("%Y-%m-%d")
-    metadata = sqlalchemy.MetaData()
+    metadata = sqlalchemy.MetaData(bind = conn)
+
+    metadata.reflect(bind = conn)
+
     
-    db_table_list = [
-        sqlalchemy.Table(
-            table_name,
-            metadata,
-            sqlalchemy.Column('retail_url',sqlalchemy.String,primary_key=True),
-            sqlalchemy.Column('data_collection_date',sqlalchemy.String),
-        ) 
-        for table_name in list_of_table_names
-    ]
-    
+    db_table_list = list(metadata.tables.keys())    
+    if not db_table_list:
+        return
     for table in db_table_list:
         delete_rows = table.delete().where(table.c.data_collection_date == today)
         conn.execute(delete_rows)
@@ -52,19 +43,12 @@ def test_push_to_db_no_today_data_tables(
         df_dict_to_append_test:dict[str,pd.DataFrame],
         df_dict_to_replace_test:dict[str,pd.DataFrame],
         gpu_data_coll_test_logger:RootLogger,
-        test_db_url:str = test_db_url,
+        test_db_url:str,
+        no_today_tables_conn: sqlalchemy.engine.Connection
     ):
     """Tests that pushing to database works when there is no data for
     'today' in the database table. Also tests if GPU units are properly prefixed
     e.g. Geforce are not prefixed by RX etc."""
-
-    
-    db_conn = sqlalchemy.create_engine(test_db_url).connect()
-    list_table_names = list(df_dict_to_append_test.keys()) +\
-            list(df_dict_to_replace_test.keys())
-        
-    delete_db_today_rows(db_conn,list_table_names)
-
     data_collection_to_db(
         test_db_url,
         df_dict_to_append_test,
@@ -72,61 +56,27 @@ def test_push_to_db_no_today_data_tables(
         gpu_data_coll_test_logger
     )
 
-
-
     gpu_of_interest_df = pd.read_sql(
         sql=sql_query_format("gpu_of_interest"),
-        con=db_conn
+        con=no_today_tables_conn
     )
-
     lowest_prices_df = pd.read_sql(
         sql=sql_query_format("lowest_prices"),
-        con=db_conn
+        con=no_today_tables_conn
     )
-
     lowest_prices_tiered = pd.read_sql(
         sql=sql_query_format("lowest_prices_tiered"),
-        con=db_conn
+        con=no_today_tables_conn
     )
 
-    gpu_of_interest_no_nan = gpu_of_interest_df.dropna()
-    
-    radeon_gpus_in_gpu_of_interest_df = gpu_of_interest_no_nan[
-        'gpu_unit_name'
-    ][gpu_of_interest_no_nan["gpu_unit_name"].str.contains('Radeon')]
-
-    geforce_gpus_in_gpu_of_interest_df = gpu_of_interest_no_nan[
-        'gpu_unit_name'
-    ][gpu_of_interest_no_nan["gpu_unit_name"].str.contains('Geforce')]
-
-    intel_gpus_in_gpu_of_interest_df = gpu_of_interest_no_nan[
-        'gpu_unit_name'
-    ][gpu_of_interest_no_nan["gpu_unit_name"].str.contains('Intel')]
-
-    
     assert len(gpu_of_interest_df) != 0
     assert len(lowest_prices_df) != 0
     assert len(lowest_prices_tiered) != 0
 
-    
-    
-    assert len(radeon_gpus_in_gpu_of_interest_df.str.contains(
-        "RTX|GTX|Arc",regex=True
-    ))!=0
-
-    assert len(geforce_gpus_in_gpu_of_interest_df.str.contains(
-            "RX|Arc",regex=True
-    ))!=0
-
-    assert len(intel_gpus_in_gpu_of_interest_df.str.contains(
-            "RX|RTX|GTX",regex=True
-    ))!=0
-    
-
 def test_push_to_db_fail_today_exists(
     df_dict_test:dict[str,pd.DataFrame],
     gpu_data_coll_test_logger: RootLogger,
-    test_db_url:str = test_db_url,
+    test_db_url:str,
 ) -> None:
     """Test for pushing to database table where 'today' data already exists"""
     list_df_name_today_exists = [
